+++ src/redux/action-creators/registry/apply.ts
@@ -1,9 +1,9 @@
 import { Nos } from 'computable/dist/types'
-import { APPLY } from '../../constants'
+import { APPLY } from '../../../constants'
 import {
   FSA,
   Applicant,
-} from '../../interfaces'
+} from '../../../interfaces'
 
 const apply = (name:string, deposit:Nos, data?:string): FSA => {
   const payload:Applicant = { name, deposit, data }
